pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes-optx-api"
version = "0.1.0"
description = "Enhanced API bridge for Hermes Agent v0.7.0+ and Hermes Workspace"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jett Optical Technologies", email = "dev@jettoptics.ai" },
]
keywords = ["hermes", "agent", "api", "bridge", "workspace"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
spacetimedb = ["requests>=2.31.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4.0"]

[project.scripts]
hermes-optx-api = "hermes_optx_api.cli:main"

[project.urls]
Homepage = "https://jettoptics.ai"
Repository = "https://github.com/jettoptx/hermes-optx-api"
Issues = "https://github.com/jettoptx/hermes-optx-api/issues"

[tool.ruff]
target-version = "py311"
line-length = 100
